fix: rename typo words and words not in Camel-Case
