+++ frontend/src/redux/actions/newsComActions.js
@@ -1,4 +1,4 @@
-import { ActionTypes } from "../contants/news-action-types";
+import { ActionTypes } from "../constants/news-action-types";
 import axios from "axios";
 export const setNewsComments = (newsComments) => {
   return {
