+++ frontend/src/redux/actions/userActions.js
@@ -1,4 +1,4 @@
-import { ActionTypes } from "../contants/user-action-types";
+import { ActionTypes } from "../constants/user-action-types";
 
 export const setUsers = (users) => {
   return {
