+++ frontend/src/redux/constants/auth-action-types.js
@@ -17,5 +17,5 @@ export const ActionTypes = {
   GOOGLE_AUTH_FAIL: "GOOGLE_AUTH_FAIL",
   FACEBOOK_AUTH_SUCCESS: "FACEBOOK_AUTH_SUCCESS",
   FACEBOOK_AUTH_FAIL: "FACEBOOK_AUTH_FAIL",
-  LOGOUT: "LOGOUT",
+  SIGNOUT: "SIGNOUT",
 };
