+++ frontend/src/redux/reducers/userReducer.js
@@ -1,4 +1,4 @@
-import { ActionTypes } from "../contants/user-action-types";
+import { ActionTypes } from "../constants/user-action-types";
 
 const initialState = {
   users: [],
@@ -8,7 +8,7 @@ export const userReducer = (state = initialState, { type, payload }) => {
   switch (type) {
     case ActionTypes.SET_USERS:
       return { ...state, users: payload };
-        case ActionTypes.REMOVE_SET_USERS:
+    case ActionTypes.REMOVE_SET_USERS:
       return {};
     default:
       return state;
